Got search working
