--- src/components/LocationCard.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function LocationCard({ name, type, dimension, residents }) {
-  return <span>todo: location</span>;
-}
