--- src/components/LocationsList.js
@@ -1,3 +1,0 @@
-import React, { useEffect, useState } from "react";
-
-export default function LocationsList() {}
