+++ src/components/WelcomePage.js
@@ -1,5 +1,4 @@
 import React from "react";
-import EnhancedSearchForm from './SearchForm';
 
 export default function WelcomePage(props) {
 
@@ -13,7 +12,6 @@ export default function WelcomePage(props) {
           alt="rick"
         />
       </header>
-      <EnhancedSearchForm/>
     </section>
   );
 }
